allow for binary streaming and storage

+++ src/main/java/com/secret/fastalign/general/SequenceId.java
@@ -51,7 +51,7 @@ public final class SequenceId
 		return this.isFwd;
 	}
 	
-	public long getHeaderId()
+	public int getHeaderId()
 	{
 		return this.id;
 	}
